# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiEnvelope_ResponseCompleteUpload_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseCompleteUpload"
              },
              {
                "type": "null"
              }
            ]
          },
          "internal_code": {
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ApiEnvelope[ResponseCompleteUpload]",
        "type": "object"
      },
      "ApiEnvelope_ResponseCreateDownload_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseCreateDownload"
              },
              {
                "type": "null"
              }
            ]
          },
          "internal_code": {
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ApiEnvelope[ResponseCreateDownload]",
        "type": "object"
      },
      "ApiEnvelope_ResponseCreateUpload_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseCreateUpload"
              },
              {
                "type": "null"
              }
            ]
          },
          "internal_code": {
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ApiEnvelope[ResponseCreateUpload]",
        "type": "object"
      },
...
      "ApiEnvelope_ResponseListDocuments_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseListDocuments"
              },
              {
                "type": "null"
              }
            ]
          },
          "internal_code": {
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ApiEnvelope[ResponseListDocuments]",
        "type": "object"
      },
...
      "DocumentDTO": {
        "properties": {
          "companyId": {
            "title": "Companyid",
            "type": "string"
          },
          "contentType": {
            "title": "Contenttype",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "companyId",
          "filename",
          "contentType",
          "size",
          "createdAt"
        ],
        "title": "DocumentDTO",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope_Literal_400__Literal__DOCUMENT_FILENAME_INVALID___": {
        "properties": {
          "code": {
            "const": 400,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "DOCUMENT_FILENAME_INVALID",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[400], Literal['DOCUMENT_FILENAME_INVALID']]",
        "type": "object"
      },
...
      "ErrorEnvelope_Literal_401__Literal__AUTH_SESSION_EXPIRED___": {
        "properties": {
          "code": {
            "const": 401,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "AUTH_SESSION_EXPIRED",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[401], Literal['AUTH_SESSION_EXPIRED']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___": {
        "properties": {
          "code": {
            "const": 403,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "REQUEST_ORIGIN_INVALID",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[403], Literal['REQUEST_ORIGIN_INVALID']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_404__Literal__COMPANY_NOT_FOUND___": {
        "properties": {
          "code": {
            "const": 404,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "COMPANY_NOT_FOUND",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[404], Literal['COMPANY_NOT_FOUND']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_404__Literal__DOCUMENT_NOT_FOUND___": {
        "properties": {
          "code": {
            "const": 404,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "DOCUMENT_NOT_FOUND",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[404], Literal['DOCUMENT_NOT_FOUND']]",
        "type": "object"
      },
...
      "ErrorEnvelope_Literal_409__Literal__DOCUMENT_SIZE_MISMATCH___": {
        "properties": {
          "code": {
            "const": 409,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "DOCUMENT_SIZE_MISMATCH",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[409], Literal['DOCUMENT_SIZE_MISMATCH']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_409__Literal__DOCUMENT_UPLOAD_INCOMPLETE___": {
        "properties": {
          "code": {
            "const": 409,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "DOCUMENT_UPLOAD_INCOMPLETE",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[409], Literal['DOCUMENT_UPLOAD_INCOMPLETE']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___": {
        "properties": {
          "code": {
            "const": 422,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "REQUEST_INVALID",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[422], Literal['REQUEST_INVALID']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___": {
        "properties": {
          "code": {
            "const": 500,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "INTERNAL_ERROR",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[500], Literal['INTERNAL_ERROR']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_503__Literal__AUTH_STORE_INVALID___": {
        "properties": {
          "code": {
            "const": 503,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "AUTH_STORE_INVALID",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[503], Literal['AUTH_STORE_INVALID']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___": {
        "properties": {
          "code": {
            "const": 503,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "AUTH_STORE_UNAVAILABLE",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[503], Literal['AUTH_STORE_UNAVAILABLE']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_503__Literal__DOCUMENT_STORAGE_UNAVAILABLE___": {
        "properties": {
          "code": {
            "const": 503,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "DOCUMENT_STORAGE_UNAVAILABLE",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[503], Literal['DOCUMENT_STORAGE_UNAVAILABLE']]",
        "type": "object"
      },
...
      "RequestCompleteUpload": {
        "properties": {
          "documentId": {
            "title": "Documentid",
            "type": "string"
          }
        },
        "required": [
          "documentId"
        ],
        "title": "RequestCompleteUpload",
        "type": "object"
      },
      "RequestCreateDownload": {
        "properties": {
          "documentId": {
            "title": "Documentid",
            "type": "string"
          }
        },
        "required": [
          "documentId"
        ],
        "title": "RequestCreateDownload",
        "type": "object"
      },
      "RequestCreateUpload": {
        "properties": {
          "companyId": {
            "title": "Companyid",
            "type": "string"
          },
          "contentType": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Contenttype",
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "size": {
            "exclusiveMinimum": 0.0,
            "maximum": 10000000.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "companyId",
          "filename",
          "contentType",
          "size"
        ],
        "title": "RequestCreateUpload",
        "type": "object"
      },
...
      "RequestListDocuments": {
        "properties": {
          "companyId": {
            "title": "Companyid",
            "type": "string"
          }
        },
        "required": [
          "companyId"
        ],
        "title": "RequestListDocuments",
        "type": "object"
      },
...
      "ResponseCompleteUpload": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentDTO"
          }
        },
        "required": [
          "document"
        ],
        "title": "ResponseCompleteUpload",
        "type": "object"
      },
      "ResponseCreateDownload": {
        "properties": {
          "expiresIn": {
            "default": 60,
            "title": "Expiresin",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ResponseCreateDownload",
        "type": "object"
      },
      "ResponseCreateUpload": {
        "properties": {
          "documentId": {
            "title": "Documentid",
            "type": "string"
          },
          "expiresIn": {
            "default": 60,
            "title": "Expiresin",
            "type": "integer"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "url",
          "headers"
        ],
        "title": "ResponseCreateUpload",
        "type": "object"
      },
...
      "ResponseListDocuments": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentDTO"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ResponseListDocuments",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Risk Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/document/complete-upload": {
      "post": {
        "operationId": "complete_upload_api_v1_document_complete_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCompleteUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ResponseCompleteUpload_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_401__Literal__AUTH_SESSION_EXPIRED___"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_404__Literal__DOCUMENT_NOT_FOUND___"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_409__Literal__DOCUMENT_UPLOAD_INCOMPLETE___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_409__Literal__DOCUMENT_SIZE_MISMATCH___"
                    }
                  ],
                  "title": "Response 409 Complete Upload Api V1 Document Complete Upload Post"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_INVALID___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__DOCUMENT_STORAGE_UNAVAILABLE___"
                    }
                  ],
                  "title": "Response 503 Complete Upload Api V1 Document Complete Upload Post"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Complete Upload",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/document/create-download": {
      "post": {
        "operationId": "create_download_api_v1_document_create_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreateDownload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ResponseCreateDownload_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_401__Literal__AUTH_SESSION_EXPIRED___"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_404__Literal__DOCUMENT_NOT_FOUND___"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_INVALID___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__DOCUMENT_STORAGE_UNAVAILABLE___"
                    }
                  ],
                  "title": "Response 503 Create Download Api V1 Document Create Download Post"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Download",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/document/create-upload": {
      "post": {
        "operationId": "create_upload_api_v1_document_create_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreateUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ResponseCreateUpload_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_400__Literal__DOCUMENT_FILENAME_INVALID___"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_401__Literal__AUTH_SESSION_EXPIRED___"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_404__Literal__COMPANY_NOT_FOUND___"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_INVALID___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__DOCUMENT_STORAGE_UNAVAILABLE___"
                    }
                  ],
                  "title": "Response 503 Create Upload Api V1 Document Create Upload Post"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Upload",
        "tags": [
          "document"
        ]
      }
    },
    "/api/v1/document/list": {
      "post": {
        "operationId": "list_documents_api_v1_document_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestListDocuments"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ResponseListDocuments_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_401__Literal__AUTH_SESSION_EXPIRED___"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_404__Literal__COMPANY_NOT_FOUND___"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_INVALID___"
                    }
                  ],
                  "title": "Response 503 List Documents Api V1 Document List Post"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Documents",
        "tags": [
          "document"
        ]
      }
    },
...
  }
}